third_party/WebKit/LayoutTests/media/controls/overlay-play-button-resizes-with-video.html

<!DOCTYPE html>
<html>
<title>Test that the overlay play button size updates with the video size</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-controls.js"></script>
<body></body>
<script>
const testCases = [
  { width: 0, height: 0, expected: 48 },
  { width: 1, height: 1, expected: 48 },
  { width: 300, height: 200, expected: 60 },
  { width: 200, height: 300, expected: 60 },
  { width: 600, height: 300, expected: 90 },
  { width: 1500, height: 1000, expected: 300 },
  { width: 700, height: 100, expected: 48 },
  { width: 700, height: 180, expected: 54 },
  { width: 700, height: 500, expected: 150 },
  { width: 700, height: 1500, expected: 210 },
];

testCases.forEach(test => {
  async_test(runTestCase, 'Video with ' + test.width + 'px width and ' + test.height + 'px height', test);
});

function runTestCase(t) {
  let test = t.properties;
  test.video = createVideo(test);
  setTimeout(t.step_func_done(() => {
    checkOverlayPlayButtonSize(test);
  }));
}

function createVideo(test) {
  let video = document.createElement('video');
  video.controls = true;
  video.preload = 'none';
  video.src = '../content/60_sec_video.webm';
  video.width = test.width;
  video.height = test.height;
  document.body.append(video);
  return video;
}

function checkOverlayPlayButtonSize(test) {
  const style = getComputedStyle(mediaControlsOverlayPlayButtonInternal(test.video));
  assert_equals(style.width, test.expected.toString() + 'px', 'video with ' + test.width +
                'px width and ' + test.height  + 'px height should have overlay play button ' +
                'width of ' + test.expected + 'px.');
  assert_equals(style.height, test.expected.toString() + 'px', 'video with ' + test.width +
                'px width and ' + test.height  + 'px height should have overlay play button ' +
                'height of ' + test.expected + 'px.');
}
</script>
</html>
